Move images controller to admin module
