+++ src/admin/admin.module.ts
@@ -1,9 +1,12 @@
 import { Module } from "@nestjs/common";
 import { ProductsController } from "./products/products.controller";
 import { SharedModule } from "../shared/shared.module";
+import { ImagesService } from "./images/images.service";
+import { ImagesController } from "./images/images.controller";
 
 @Module({
   imports: [SharedModule],
-  controllers: [ProductsController],
+  controllers: [ProductsController, ImagesController],
+  providers: [ImagesService],
 })
 export class AdminModule {}
