+++ src/admin/images/dto/CloudinaryInfo.dto.ts
