+++ src/admin/images/dto/UploadImage.dto.ts
