+++ src/admin/images/images.controller.spec.ts
@@ -1,20 +1,18 @@
-import { Test, TestingModule } from "@nestjs/testing";
-import { ImagesController } from "./images.controller";
-import { ImagesService } from "./images.service";
+import { Test, TestingModule } from '@nestjs/testing';
+import { ImagesController } from './images.controller';
 
-describe("Images Controller", () => {
+describe('Images Controller', () => {
   let controller: ImagesController;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
       controllers: [ImagesController],
-      providers: [ImagesService],
     }).compile();
 
     controller = module.get<ImagesController>(ImagesController);
   });
 
-  it("should be defined", () => {
+  it('should be defined', () => {
     expect(controller).toBeDefined();
   });
 });
