+++ src/admin/images/images.service.spec.ts
@@ -1,7 +1,7 @@
-import { Test, TestingModule } from "@nestjs/testing";
-import { ImagesService } from "./images.service";
+import { Test, TestingModule } from '@nestjs/testing';
+import { ImagesService } from './images.service';
 
-describe("ImagesService", () => {
+describe('ImagesService', () => {
   let service: ImagesService;
 
   beforeEach(async () => {
@@ -12,7 +12,7 @@ describe("ImagesService", () => {
     service = module.get<ImagesService>(ImagesService);
   });
 
-  it("should be defined", () => {
+  it('should be defined', () => {
     expect(service).toBeDefined();
   });
 });
